+++ custom_components/solarcharger/config/config_subentry_charger.py
@@ -54,7 +54,7 @@ from ..const import (
 )
 from ..exceptions.validation_exception import ValidationExceptionError
 from ..helpers.utils import compose_subdomain
-from .config_options_flow import reset_api_entities
+from .config_options_flow import process_api_config
 from .config_utils import get_subentry_id
 
 # ----------------------------------------------------------------------------
@@ -159,7 +159,7 @@ class AddChargerSubEntryFlowHandler(ConfigSubentryFlow):
         data: dict[str, Any] = {
             OPTION_CHARGER_NAME: device_name,
         }
-        reset_api_entities(config_entry, subentry_unique_id, data)
+        process_api_config(config_entry, subentry_unique_id, data, is_init_all=True)
 
         # Use | (union) to replace or add key:data pair.
         self.hass.config_entries.async_update_entry(