+++ custom_components/solarcharger/number.py
@@ -51,7 +51,7 @@ from .const import (
     NUMBER_DEFAULT_CHARGE_LIMIT_THURSDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_TUESDAY,
     NUMBER_DEFAULT_CHARGE_LIMIT_WEDNESDAY,
-    NUMBER_MAX_CONSUMED_ENERGY,
+    NUMBER_MAX_CONSUMED_ENERGY_LIMIT,
     NUMBER_OCPP_PROFILE_ID,
     NUMBER_OCPP_PROFILE_STACK_LEVEL,
     NUMBER_POWER_MONITOR_DURATION,
@@ -285,10 +285,10 @@ CONFIG_NUMBER_LIST: tuple[
         ),
     ),
     (
-        NUMBER_MAX_CONSUMED_ENERGY,
+        NUMBER_MAX_CONSUMED_ENERGY_LIMIT,
         SolarChargerEntityType.TYPE_LOCAL,
         NumberEntityDescription(
-            key=NUMBER_MAX_CONSUMED_ENERGY,
+            key=NUMBER_MAX_CONSUMED_ENERGY_LIMIT,
             entity_category=EntityCategory.CONFIG,
             device_class=NumberDeviceClass.ENERGY,
             native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,